refactor: consolidate prompts, parseScore, and fix minor quality issues
- Move TOPIC_RANKING_SYSTEM_PROMPT and getTopicComparisonPrompt to
  src/lib/prompts.ts (canonical prompt location)
- Move parseScore to src/lib/score-utils.ts (reusable Drizzle decimal
  conversion)
- Extract LOOKBACK_DAYS constant in rank-episode-topics.ts
- Remove unnecessary optional markers on RecommendedEpisodeDTO fields
- Simplify rankMap construction (remove redundant wrapper object)
- Batch rank writes with Promise.all (eliminates sequential awaits)
- Fix what-comment to explain why (separate query avoids LIMIT
  expansion)
